requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-allauth==0.53.1
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.1
django-countries==7.5.1
django-environ==0.10.0
django-import-export==3.1.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-redis==5.2.0
django-tinymce==3.6.1
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-http-client==3.3.7
python3-openid==3.2.0
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tablib==3.4.0
typing-extensions==4.5.0
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
